# docs/swagger.yaml
basePath: /api
definitions:
...
  handlers.AbsensiResponse:
    properties:
      created_at:
        type: string
      deskripsi:
        type: string
      id_absen:
        type: integer
      id_jadwal:
        type: integer
      nis:
        type: string
      status:
        type: string
      tanggal:
        type: string
    type: object
...
  handlers.AppliedFilters:
    properties:
      jk:
        type: string
      jurusan:
        type: string
      kelas:
        type: string
      search:
        type: string
      sort_by:
        type: string
      sort_dir:
        type: string
    type: object
...
  handlers.CreateAbsensiRequest:
    properties:
      deskripsi:
        type: string
      id_jadwal:
        type: integer
      status:
        enum:
        - hadir
        - izin
        - sakit
        - alpha
        type: string
      tanggal:
        type: string
    required:
    - id_jadwal
    - status
    - tanggal
    type: object
...
  handlers.PaginationMeta:
    properties:
      page:
        type: integer
      page_size:
        type: integer
      total_items:
        type: integer
      total_pages:
        type: integer
    type: object
...
  handlers.SiswaCreateResponse:
    properties:
      data:
        $ref: '#/definitions/models.Siswa'
      message:
        type: string
    type: object
  handlers.SiswaDeleteResponse:
    properties:
      message:
        type: string
    type: object
  handlers.SiswaDetailResponse:
    properties:
      data:
        $ref: '#/definitions/models.Siswa'
      message:
        type: string
    type: object
  handlers.SiswaErrorResponse:
    properties:
      error: {}
      message:
        type: string
    type: object
...
  handlers.SiswaListPaginatedResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/models.Siswa'
        type: array
      filters:
        $ref: '#/definitions/handlers.AppliedFilters'
      message:
        type: string
      pagination:
        $ref: '#/definitions/handlers.PaginationMeta'
    type: object
  handlers.SiswaUpdateResponse:
    properties:
      data:
        $ref: '#/definitions/models.Siswa'
      message:
        type: string
    type: object
...
  models.Siswa:
    properties:
      jk:
        type: string
      jurusan:
        type: string
      kelas:
        type: string
      nama_siswa:
        type: string
      nis:
        type: string
    type: object
...
info:
  contact:
    name: API Support
  description: API untuk sistem pencatatan absensi sholat siswa
  license:
    name: MIT
  termsOfService: http://swagger.io/terms/
  title: API Sistem Absensi Sholat
  version: 1.0.0
paths:
...
  /siswa:
    get:
      consumes:
      - application/json
      description: Mengambil daftar siswa dengan dukungan filter, pencarian, sorting,
        dan pagination
      parameters:
      - description: Cari berdasarkan NIS atau nama siswa
        in: query
        name: search
        type: string
      - description: Filter berdasarkan kelas (exact match)
        in: query
        name: kelas
        type: string
      - description: Filter berdasarkan jurusan (exact match)
        in: query
        name: jurusan
        type: string
      - description: Filter berdasarkan jenis kelamin (Laki-laki/Perempuan)
        in: query
        name: jk
        type: string
      - description: 'Nomor halaman (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Jumlah item per halaman (default: 20, max: 100)'
        in: query
        name: page_size
        type: integer
      - description: 'Urutkan berdasarkan: nis, nama_siswa, kelas, jurusan (default:
          nama_siswa)'
        in: query
        name: sort_by
        type: string
      - description: 'Arah pengurutan: asc, desc (default: asc)'
        in: query
        name: sort_dir
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Data siswa berhasil diambil dengan metadata pagination
          schema:
            $ref: '#/definitions/handlers.SiswaListPaginatedResponse'
        "500":
          description: Kesalahan server internal - Gagal query database
          schema:
            $ref: '#/definitions/handlers.SiswaErrorResponse'
      summary: Ambil semua data siswa dengan filter dan pagination
      tags:
      - siswa
    post:
      consumes:
      - application/json
      description: Menambahkan siswa baru ke dalam sistem. NIS harus unik dan belum
        terdaftar. Semua field (NIS, nama_siswa, jk) wajib diisi
      parameters:
      - description: Data siswa baru - NIS (unique), nama_siswa, jk (Laki-laki/Perempuan),
          jurusan (optional), kelas (optional)
        in: body
        name: student
        required: true
        schema:
          $ref: '#/definitions/models.Siswa'
      produces:
      - application/json
      responses:
        "201":
          description: Siswa berhasil ditambahkan - Mengembalikan data siswa yang
            baru dibuat
          schema:
            $ref: '#/definitions/handlers.SiswaCreateResponse'
        "400":
          description: Permintaan tidak valid - Format JSON salah, field required
            kosong, atau duplicate NIS
          schema:
            $ref: '#/definitions/handlers.SiswaErrorResponse'
        "500":
          description: Kesalahan server internal - Database constraint violation atau
            connection error
          schema:
            $ref: '#/definitions/handlers.SiswaErrorResponse'
      summary: Tambah data siswa baru
      tags:
      - siswa
  /siswa/{nis}:
    delete:
      consumes:
      - application/json
      description: Menghapus siswa dari sistem secara permanen. Operasi ini akan menghapus
        semua data terkait siswa termasuk akun login dan absensi. Hati-hati karena
        operasi ini tidak dapat dibatalkan
      parameters:
      - description: NIS Siswa yang akan dihapus
        in: path
        name: nis
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Siswa berhasil dihapus - Data siswa telah dihapus dari sistem
          schema:
            $ref: '#/definitions/handlers.SiswaDeleteResponse'
        "400":
          description: Permintaan tidak valid - NIS kosong atau tidak diberikan
          schema:
            $ref: '#/definitions/handlers.SiswaErrorResponse'
        "404":
          description: Siswa tidak ditemukan - NIS tidak ada dalam database
          schema:
            $ref: '#/definitions/handlers.SiswaErrorResponse'
        "500":
          description: Kesalahan server internal - Database error atau cascade delete
            error
          schema:
            $ref: '#/definitions/handlers.SiswaErrorResponse'
      summary: Hapus data siswa
      tags:
      - siswa
    get:
      consumes:
      - application/json
      description: Mengambil detail siswa spesifik berdasarkan nomor induk siswa (NIS).
        Mengembalikan profil lengkap siswa termasuk nama, jenis kelamin, jurusan,
        dan kelas
      parameters:
      - description: NIS Siswa - nomor induk siswa 4-20 digit
        in: path
        name: nis
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Data siswa ditemukan dan berhasil diambil
          schema:
            $ref: '#/definitions/handlers.SiswaDetailResponse'
        "400":
          description: NIS tidak valid atau kosong - parameter path tidak diberikan
          schema:
            $ref: '#/definitions/handlers.SiswaErrorResponse'
        "404":
          description: Siswa tidak ditemukan - NIS tidak ada dalam database
          schema:
            $ref: '#/definitions/handlers.SiswaErrorResponse'
        "500":
          description: Kesalahan server internal - Database connection error atau
            query error
          schema:
            $ref: '#/definitions/handlers.SiswaErrorResponse'
      summary: Ambil data siswa berdasarkan NIS
      tags:
      - siswa
    put:
      consumes:
      - application/json
      description: Memperbarui informasi siswa yang sudah terdaftar. NIS tidak dapat
        diubah (primary key). Hanya nama_siswa, jk, jurusan, dan kelas yang dapat
        diperbarui
      parameters:
      - description: NIS Siswa yang akan diupdate
        in: path
        name: nis
        required: true
        type: string
      - description: Data siswa yang diperbarui - nama_siswa, jk, jurusan, kelas dapat
          diubah
        in: body
        name: student
        required: true
        schema:
          $ref: '#/definitions/models.Siswa'
      produces:
      - application/json
      responses:
        "200":
          description: Siswa berhasil diperbarui - Mengembalikan data siswa terbaru
          schema:
            $ref: '#/definitions/handlers.SiswaUpdateResponse'
        "400":
          description: Permintaan tidak valid - Format JSON salah atau NIS kosong
          schema:
            $ref: '#/definitions/handlers.SiswaErrorResponse'
        "404":
          description: Siswa tidak ditemukan - NIS tidak ada dalam database
          schema:
            $ref: '#/definitions/handlers.SiswaErrorResponse'
        "500":
          description: Kesalahan server internal - Database error atau constraint
            violation
          schema:
            $ref: '#/definitions/handlers.SiswaErrorResponse'
      summary: Ubah data siswa
      tags:
      - siswa
  /siswa/{nis}/absensi:
    post:
      consumes:
      - application/json
      description: 'Siswa membuat pencatatan absensi (kehadiran/ketidakhadiran) untuk
        sesi sholat. Status yang valid adalah: hadir, izin, sakit, alpha'
      parameters:
      - description: NIS Siswa - nomor induk siswa
        in: path
        name: nis
        required: true
        type: string
      - description: Data absensi siswa
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateAbsensiRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Absensi berhasil dibuat
          schema:
            $ref: '#/definitions/handlers.AbsensiResponse'
        "400":
          description: Request tidak valid atau data tidak lengkap
          schema:
            $ref: '#/definitions/handlers.SiswaErrorResponse'
        "404":
          description: Siswa atau jadwal sholat tidak ditemukan
          schema:
            $ref: '#/definitions/handlers.SiswaErrorResponse'
        "500":
          description: Kesalahan server internal
          schema:
            $ref: '#/definitions/handlers.SiswaErrorResponse'
      summary: Buat absensi sholat baru untuk siswa
      tags:
      - absensi
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"